+++ src/App/Fossa/FossaAPIV1.hs
@@ -198,6 +198,8 @@ mkMetadataOpts ProjectMetadata{..} projectName = mconcat $ catMaybes maybes
       , ("link" =:) <$> projectLink
       , ("team" =:) <$> projectTeam
       , ("policy" =:) <$> projectPolicy
+      , ("releaseGroup" =:) <$> projectReleaseGroupName
+      , ("releaseGroupRelease" =:) <$> projectReleaseGroupRelease
       , ("title" =:) <$> title
       ]